extract complex methods in /glyph
