+++ src/commands/5/glyph.ts
@@ -1,9 +1,11 @@
-import { ApplicationCommandOptionType, ApplicationCommandType, AttachmentBuilder, ChatInputCommandInteraction, EmbedBuilder, User } from "discord.js";
-import { GlyphEmbedGetter, basicGlyphs, specialGlyphs } from "../../utils/databases/glyphs";
-import { authorTitle, isHelper } from "../../functions/Misc";
+import { ApplicationCommandOptionType, ApplicationCommandType, ChatInputCommandInteraction } from "discord.js";
+import { basicGlyphs, specialGlyphs } from "../../utils/databases/glyphs";
 import { Command } from "../../command";
 import { GlyphInfo } from "../../utils/types";
 import config from "../../config.json";
+import { effectGlyphSubcommand } from "./glyph/effect";
+import { isHelper } from "../../functions/Misc";
+import { sacrificeGlyphSubcommand } from "./glyph/sacrifice";
 import { utilsGlyphSubcommand } from "./glyph/utils";
 
 function getEffectChoices(): { name: string, value: string, type: any }[] {
@@ -275,8 +277,6 @@ export const glyph: Command = {
       return;
     }
 
-    const user: User = interaction.member === null ? interaction.user : interaction.member.user as User;
-
     const isADServer: boolean = (interaction.guildId === config.ids.AD.serverID);
 
     if (interaction.options.getSubcommandGroup() !== null) {
@@ -294,40 +294,9 @@ export const glyph: Command = {
       await interaction.reply({ content, ephemeral: !isHelper(interaction) });
 
     } else if (type === "effect") {
-
-      const glyphName: string = interaction.options.getString("glyph") as string;
-      const altered: boolean = interaction.options.getBoolean("altered") as boolean;
-      const hasAlteredImage: boolean = ["dilation", "effarig", "infinity", "power", "replication", "time"].includes(glyphName);
-
-      let picture = new AttachmentBuilder(`src/images/glyphs/${glyphName}.png`);
-      if (altered && hasAlteredImage) picture = new AttachmentBuilder(`src/images/glyphs/altered/${glyphName}_altered.png`);
-      else if (altered) picture = new AttachmentBuilder(`src/images/glyphs/altered/none_altered.png`);
-
-      const glyphRequested = basicGlyphs[glyphName] ?? specialGlyphs[glyphName];
-
-      let imageLink: string = `attachment://${glyphName}.png`;
-      if (altered && hasAlteredImage) imageLink = `attachment://${glyphName}_altered.png`;
-      else if (altered) imageLink = `attachment://none_altered.png`;
-
-      const embed: EmbedBuilder = GlyphEmbedGetter(glyphRequested, isADServer, altered, false);
-      embed.setAuthor({ name: authorTitle(interaction), iconURL: user.displayAvatarURL() })
-        .setThumbnail(imageLink);
-
-      await interaction.reply({ embeds: [embed], files: [picture], ephemeral: !isHelper(interaction) });
+      await effectGlyphSubcommand(interaction);
     } else {
-      const glyphName: string = interaction.options.getString("glyph") as string;
-      const hasSacrificeImage: boolean = ["dilation", "effarig", "infinity", "power", "reality", "replication", "time"].includes(glyphName);
-
-      let picture: AttachmentBuilder = new AttachmentBuilder(`src/images/glyphs/sacrificed/${glyphName}.png`);
-      if (!hasSacrificeImage) picture = new AttachmentBuilder(`src/images/glyphs/${glyphName}.png`);
-      const imageLink: string = `attachment://${glyphName}.png`;
-      const glyphRequested = basicGlyphs[glyphName] ?? specialGlyphs[glyphName];
-
-      const embed: EmbedBuilder = GlyphEmbedGetter(glyphRequested, isADServer, false, true);
-      embed.setAuthor({ name: authorTitle(interaction), iconURL: user.displayAvatarURL() })
-        .setThumbnail(imageLink);
-
-      await interaction.reply({ embeds: [embed], files: [picture], ephemeral: !isHelper(interaction) });
+      await sacrificeGlyphSubcommand(interaction);
     }
   }
 };
